Add: Decisions to Navigation
